Waves!
Added the wave class. Also had to partially re-implement the player
class in order to get the wave class to work - the player class has
large amounts of I/O which I haven't figured out how I'm going to do in
C# yet, so it's all been commented out.

The rotation method seems to cut the sprite strangely, perhaps it does
not get moved back properly when it is moved to be rotated centrally
using the graphics object. The image transparency also needs to be
looked at.